refactor: Remove scrollbar-gutter property from styleUtils
The scrollbar-gutter property was causing issues with the scrollbars in the application. This commit removes the property to fix the problem.

Refactor the styleUtils module in the shared/utils directory to remove the scrollbar-gutter property from the overflowYAuto style. This will ensure that the scrollbars in the application function correctly.